+++ src/mapbox/mapbox.js
@@ -40,6 +40,7 @@ const propTypes = {
   onLoad: PropTypes.func, /** The onLoad callback for the map */
   onError: PropTypes.func, /** The onError callback for the map */
   reuseMaps: PropTypes.bool,
+  reuseMap: PropTypes.bool,
   transformRequest: PropTypes.func, /** The transformRequest callback for the map */
 
   mapStyle: PropTypes.string, /** The Mapbox style. A string url to a MapboxGL style */
@@ -66,6 +67,7 @@ const defaultProps = {
   onLoad: noop,
   onError: noop,
   reuseMaps: false,
+  reuseMap: false,
   transformRequest: null,
 
   mapStyle: 'mapbox://styles/mapbox/light-v8',
@@ -165,7 +167,7 @@ export default class Mapbox {
 
   _create(props) {
     // Reuse a saved map, if available
-    if (props.reuseMaps && Mapbox.savedMap) {
+    if ((props.reuseMaps || props.reuseMap) && Mapbox.savedMap) {
       this._map = this.map = Mapbox.savedMap;
       // When reusing the saved map, we need to reparent the map(canvas) and other child nodes
       // intoto the new container from the props.
@@ -179,6 +181,12 @@ export default class Mapbox {
       // Step2: replace the internal container with new container from the react component
       this._map._container = newContainer;
       Mapbox.savedMap = null;
+
+      // Update style
+      if (props.mapStyle) {
+        this._map.setStyle(props.mapStyle);
+      }
+
       // TODO - need to call onload again, need to track with Promise?
       props.onLoad();
     } else {