+++ docs/whats-new.md
@@ -1,3 +1,4 @@
+
 # react-map-gl v3.2
 
 Realease date: January, 2018