+++ docs/whats-new.md
@@ -59,11 +59,11 @@ Release date: July 27th, 2017
 
 ### [InteractiveMap (New, MapGL replacement)](/#/Documentation/api-reference/interactive-map)
 
-This is a wrapper on top of `StaticMap`. It takes all the props of `StaticMap` and additional ones such as `onViewportChange`, `scrollZoom`, `dragRotate`, etc. to control interactivity on the map. See [Source Code](https://github.com/uber/react-map-gl/blob/master/src/components/interactive-map.js) for more information.
+This is a wrapper on top of `StaticMap`. It takes all the props of `StaticMap` and additional ones such as `onViewportChange`, `scrollZoom`, `dragRotate`, etc. to control interactivity on the map. See [Source Code](https://github.com/uber/react-map-gl/blob/3.3-release/src/components/interactive-map.js) for more information.
 
 ### [StaticMap (New)](/#/Documentation/api-reference/static-map)
 
-This is the React wrapper around `Mapbox GL JS` and takes in viewport properties such as `width`, `height`, `latitude`, `longitude`. Style diffing and updating logic also live here. See [Source Code](https://github.com/uber/react-map-gl/blob/master/src/components/static-map.js) for more information.
+This is the React wrapper around `Mapbox GL JS` and takes in viewport properties such as `width`, `height`, `latitude`, `longitude`. Style diffing and updating logic also live here. See [Source Code](https://github.com/uber/react-map-gl/blob/3.3-release/src/components/static-map.js) for more information.
 
 ### Overlays